Call mypy with newly added parameter `flush_errors`
Unfortunately the target function does not have any docstring in mypy,
but it is typed as Optional[] which means the function should accept
None as a value. Doing so seems to suppress issues reported in #6.

Quick skim of the culprit PR in mypy
(https://github.com/python/mypy/pull/4396/files) shows the code really
counts with None being passed and adjusting appropriately, citing:

> If we were not given a flush_errors, we use one that will populate
> those fields for callers that want the traditional API.